Implement Structural Metadata Editor
This feature works with the theory that structural metadata can be
abstracted into a hash that looks like the following:

```
{
  "nodes": [
    {
      "label": "Chapter 1",
      "nodes": [
        {
          "proxy": "mcvaa9zja"
        }
      ]
    }
  ]
}
```

This hash is created by the frontend with `StructureParser`, is
converted into an ordered graph by `LogicalOrder`, and can be created
from an ordered graph by `LogicalOrderGraph`.

This hash is also used to create IIIF Manifest Ranges for the viewer.

Closes #225, closes #18.